Implement "Hello, world" functionality
The bot now successfully connects to Rizon, joins #ircbottestrelam,
and responds to messages beginning with "!test" with the sentence
"The quick brown fox jumps over the lazy dog." It handles PING/PONG
events, but not VERSION requests. It ignores all messages for the
first thirty seconds of connection, instead of parsing the welcome
messages.