Set up framework for basic command handling
The new file codelist.go contains a map from IRC commands such as
"PING" or "PRIVMSG" to a function that handles each command. The only
fully-implemented commands so far are responses to PING and VERSION.
Any message containing the command PRIVMSG will echo the in-memory
representation of the message to stdout. All other commands are
not-yet-implemented. The file parseMessage_test.go provides basic
testing functionality for these commands.

The logic in main.go was changed to accommodate this new, more
extensible setup.